requirements.txt
Flask==3.1.3
gunicorn==23.0.0
Pillow==12.1.1
pymongo==4.16.0
cloudinary==1.44.1
openpyxl==3.1.5
qrcode==8.2
python-dotenv==1.2.1
requests==2.32.5
dnspython==2.8.0
redis==7.2.1
certifi==2024.12.14
flask-talisman==1.1.0
flask-cors==4.0.0
celery==5.4.0
flask-limiter==3.5.0
pybreaker==1.0.2
flask-session==0.6.0
cachelib==0.10.2
opencv-python==4.10.0.84
